+++ src/imio/dashboard/tests/test_vocabulary.py
@@ -6,6 +6,7 @@ from zope.lifecycleevent import ObjectModifiedEvent
 from zope.schema.interfaces import IVocabularyFactory
 from Products.CMFCore.utils import getToolByName
 from plone.app.testing import login
+from plone.app.testing import TEST_USER_NAME
 from plone import api
 from imio.dashboard.testing import IntegrationTestCase
 from collective.behavior.talcondition.interfaces import ITALConditionable
@@ -32,6 +33,12 @@ class TestConditionAwareVocabulary(IntegrationTestCase):
         """This vocabulary is condition aware, it means
            that it will take into account condition defined in the
            'tal_condition' field added by ITALConditionable."""
+        # add on non Manager user
+        api.user.create(
+            username='user_not_manager',
+            password='user_not_manager',
+            email="imio@dashboard.org",
+            roles=['Member'])
         self.assertTrue(ITALConditionable.providedBy(self.dashboardcollection))
         factory = queryUtility(IVocabularyFactory, u'imio.dashboard.conditionawarecollectionvocabulary')
         # for now, no condition defined on the collection so it is in the vocabulary
@@ -41,10 +48,16 @@ class TestConditionAwareVocabulary(IntegrationTestCase):
         # now define a condition and by pass for Manager
         self.dashboardcollection.tal_condition = u'python:False'
         self.dashboardcollection.roles_bypassing_talcondition = [u"Manager"]
+        notify(ObjectModifiedEvent(self.dashboardcollection))
         # No more listed except for Manager
         vocab = factory(self.portal)
         self.assertTrue(self.dashboardcollection.UID() in [term.token for term in vocab])
+        login(self.portal, 'user_not_manager')
+        # cache is user aware
+        vocab = factory(self.portal)
+        self.assertFalse(self.dashboardcollection.UID() in [term.token for term in vocab])
         # Now, desactivate by pass for manager
+        login(self.portal, TEST_USER_NAME)
         self.dashboardcollection.roles_bypassing_talcondition = []
         # ObjectModified event on DashboardCollection invalidate the vocabulary caching
         notify(ObjectModifiedEvent(self.dashboardcollection))