+++ CHANGES.rst
@@ -6,8 +6,8 @@ Changelog
 ----------------
 
 - Use `ram.cache` for the `imio.dashboard.conditionawarecollectionvocabulary`
-  vocabulary.  This is closest faceted context relative and is invalidated when
-  a `DashboardCollection` is modified.
+  vocabulary.  This is user and closest faceted context relative and is
+  invalidated when a `DashboardCollection` is modified.
   [gbastien]
 
 1.3 (2018-01-06)